docs: v2 架构决策——转投 libghostty 嵌入（cmux 路线）
- PLAN-V2.md：q0-q5 阶段计划（q0 含嵌入 API 能力审计硬门禁：hit 数据源/层合成）
- STACK.md 顶部 v2 修订：引擎/渲染/PTY/配置全套用 Ghostty，ninja=ADE 插件宿主
- 保留：protocol/preview/theme 三 crate、监督器、面板；退役：自研引擎层（q4 tag）
- 2026-08-30 用户产品决策（三选一选 B）

+++ crates/ninja/src/font.rs
@@ -494,9 +494,9 @@ impl Font {
 fn measure(font: &CTFont, size_pt: f64) -> Metrics {
     // SAFETY: font/缓冲参数布局正确，单字形测量。
     let (ascent, descent, leading, advance) = unsafe {
-        let ascent = font.ascent() as f64;
-        let descent = font.descent() as f64;
-        let leading = font.leading() as f64;
+        let ascent = font.ascent();
+        let descent = font.descent();
+        let leading = font.leading();
         let mut chars: [u16; 1] = [b'M' as u16];
         let mut glyphs: [u16; 1] = [0];
         let ok = font.glyphs_for_characters(
